dom/security/test/csp/test_security_policy_violation_event.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta http-equiv="Content-Security-Policy" content="img-src 'none'">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script>
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({
  set: [
    ["security.csp.enable_violation_events", true]
  ]
});
document.addEventListener("securitypolicyviolation", (e) => {
  SimpleTest.is(e.blockedURI, "http://mochi.test:8888/foo/bar.jpg", "blockedURI");
  SimpleTest.is(e.violatedDirective, "img-src", "violatedDirective")
  SimpleTest.is(e.originalPolicy, "img-src 'none'", "originalPolicy");
  SimpleTest.finish();
});
</script>
<img src="http://mochi.test:8888/foo/bar.jpg">
